iot-portal/requirements.txt
# INSA Advanced IIoT Platform - Production Dependencies
# Version: 2.0
# Updated: October 28, 2025

# Core Flask Framework
Flask==3.1.2
Flask-CORS==5.0.0
Flask-SocketIO==5.5.1
Flask-JWT-Extended==4.7.1
Flasgger==0.9.7.1

# Database & Caching
psycopg2-binary==2.9.10
redis==5.2.0

# Data Processing
pandas==2.3.2
numpy==2.3.3
scipy==1.15.1

# Analytics & Statistics
scikit-learn==1.6.1

# Communication Protocols
paho-mqtt==2.1.0
requests==2.32.3

# Security & Rate Limiting
Flask-Limiter==3.8.0
cryptography==44.0.0

# Utilities
python-dotenv==1.0.1
python-socketio==5.12.1
